requirements.txt
tqdm
GitPython
pyyaml